Make cookieID a 32-bit value instead of 16-bit

+++ xwords4/relay/xwrelay_priv.h
@@ -6,7 +6,7 @@
 #include <time.h>
 
 typedef unsigned short HostID;
-typedef unsigned short CookieID; /* stands in for string after connection established */
+typedef unsigned long CookieID; /* stands in for string after connection established */
 
 void logf( const char* format, ... );
